requirements.txt
anyio==3.6.2
bcrypt==4.1.1
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.2.0
fastapi==0.95.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
fpdf==1.7.2
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
Pillow==9.4.0
pluggy==1.3.0
psycopg2-binary==2.9.9
pydantic==1.10.7
PyJWT==2.8.0
pytesseract==0.3.10
pytest==7.4.3
python-dotenv==1.0.0
PyYAML==6.0
reportlab==4.0.4
requests==2.28.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==10.4
Werkzeug==2.2.3
